requirements.txt
affine>=2.3,<3.0
attrs>=25.4,<26.0
certifi>=2026.1.4,<2027
click>=8.3,<9.0
click-plugins>=1.1.1,<2.0
cligj>=0.7.2,<0.9
contourpy>=1.3,<2.0
cycler>=0.12,<0.13
filelock>=3.20,<4.0
fonttools>=4.61,<5.0
fsspec>=2025.12,<2026
imageio>=2.37,<3.0
Jinja2>=3.1.6,<4.0
kiwisolver>=1.4.9,<2.0
lazy_loader>=0.4,<0.5
MarkupSafe>=3.0.3,<4.0
matplotlib>=3.10.8,<4.0
mpmath>=1.3,<2.0
networkx>=3.6,<4.0
numpy>=2.3.5,<3.0
pillow>=12.0,<13.0
pyparsing>=3.3.1,<4.0
python-dateutil>=2.9.0.post0,<3.0
rasterio>=1.4.4,<2.0
scikit-image>=0.26,<0.27
scipy>=1.16.3,<2.0
setuptools>=80.9,<81.0
six>=1.17,<2.0
sympy>=1.14,<2.0
tifffile>=2025.10.16,<2026
torch>=2.9.1,<3.0
torchvision>=0.24.1,<0.25
tqdm>=4.67.1,<5.0
triton>=3.5.1,<4.0
typing_extensions>=4.15,<5.0
packaging>=25.0,<26.0